+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+### 4.0.0
+* Remove Nosto category personalization feature from this extension
+* Remove product indexer
+* Fix an issue where the product's category was encoded twice
+
 ### 3.11.4
 * Fix category tagging not being rendered on product detail page
 * Fix an issue when rendering category tagging for categories in which parents are hidden